+++ vue-src/index.ts
@@ -1,0 +1,18 @@
+export { useGenerate } from "./useGenerate.js";
+export { useGenerateSvg } from "./useGenerateSvg.js";
+export { useGenerateModules } from "./useGenerateModules.js";
+export { useDecode } from "./useDecode.js";
+export { useQrScanner } from "./useQrScanner.js";
+export type { UseQrScannerOptions } from "./useQrScanner.js";
+export type { MaybeRefOptions } from "./shared.js";
+
+export type {
+  DecodedQr,
+  GenerateOptions,
+  HqrError,
+  HqrErrorCode,
+  QrEcc,
+  QrModules,
+  QrSizeMode,
+  SvgOptions,
+} from "../index";